Seat Booking System
First the user chooses if they want club or executive seats. Proceeding, an matrix representing seats is printed. The user enters the number of seats as well as the seats they want to book. Following this, a matrix depicting booked seats (represented by 0) is printed along with the money the user needs to pay.